+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 予約取消アクションのIObjectを最適化
+- イベントオファー承認アクションのobject.eventを最適化
 
 ### Deprecated